Refactor KDoc to centralize auto-renaming topics across modules
