+++ dataframe-csv/src/main/kotlin/org/jetbrains/kotlinx/dataframe/documentationCsv/CommonReadDelimDocs.kt
@@ -8,14 +8,13 @@ import org.jetbrains.kotlinx.dataframe.io.ColType
 import java.io.File
 import java.io.InputStream
 import java.net.URL
-import org.jetbrains.kotlinx.dataframe.documentation.AutoRenameInputSnippet
 
 /**
  * ### Read $[FILE_TYPE_TITLE] $[DATA_TITLE] to [DataFrame]
  *
  * Reads any $[FILE_TYPE] $[DATA] to a [DataFrame][DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * Parameters you can use to customize the reading process include, for instance, \[delimiter\],
  * \[header\], \[colTypes\], \[readLines\], and \[parserOptions\].
