+++ dataframe-excel/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/xlsx.kt
@@ -32,9 +32,9 @@ import org.jetbrains.kotlinx.dataframe.api.forEach
 import org.jetbrains.kotlinx.dataframe.api.select
 import org.jetbrains.kotlinx.dataframe.codeGen.AbstractDefaultReadMethod
 import org.jetbrains.kotlinx.dataframe.codeGen.DefaultReadDfMethod
-import org.jetbrains.kotlinx.dataframe.documentation.AutoRenameInputSnippet
 import org.jetbrains.kotlinx.dataframe.exceptions.DuplicateColumnNamesException
 import org.jetbrains.kotlinx.dataframe.impl.ColumnNameGenerator
+import org.jetbrains.kotlinx.dataframe.io.documentation.AutoRenameInputSnippet
 import org.jetbrains.kotlinx.dataframe.util.DF_READ_EXCEL
 import org.jetbrains.kotlinx.dataframe.util.READ_EXCEL_NAME_REPAIR
 import java.io.File
@@ -96,7 +96,7 @@ private fun setWorkbookTempDirectory() {
 /**
  * Reads an Excel sheet from the given [URL] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * {@include [AutoRenameInputSnippet]}
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -139,7 +139,7 @@ public fun DataFrame.Companion.readExcel(
 /**
  * Reads an Excel sheet from the given [File] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -177,7 +177,7 @@ public fun DataFrame.Companion.readExcel(
 /**
  * Reads an Excel sheet from the given [Path] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -223,7 +223,7 @@ public fun DataFrame.Companion.readExcel(
 /**
  * Reads an Excel sheet from the given file or URL into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -261,7 +261,7 @@ public fun DataFrame.Companion.readExcel(
 /**
  * Reads an Excel sheet from the given [InputStream] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -304,7 +304,7 @@ public fun DataFrame.Companion.readExcel(
 /**
  * Reads an Excel sheet from [Workbook] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheetName sheet to read. By default, the first sheet in the document
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
@@ -363,7 +363,7 @@ public class FormattingOptions(range: String, public val formatter: DataFormatte
 /**
  * Reads Excel [Sheet] into a [DataFrame].
  *
- * @include [AutoRenameInputSnippet].
+ * @include [AutoRenameInputSnippet]
  *
  * @param sheet sheet to read.
  * @param columns comma separated list of Excel column letters and column ranges (e.g. “A:E” or “A,C,E:F”)
