+++ dataframe-jdbc/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     with(convention.plugins) {
         alias(kotlinJvm8)
         alias(buildConfig)
+        alias(kodex)
     }
     with(libs.plugins) {
         alias(publisher)
